Fix test_compile cases

+++ test_compile/compile_dyn_dbn.cpp
@@ -13,7 +13,7 @@ template<typename DBN>
 void test_dbn(DBN& dbn){
     dbn->display();
 
-    std::vector<etl::dyn_vector<double>> images;
+    std::vector<etl::dyn_vector<float>> images;
 
     dbn->pretrain(images, 10);
 }
